Moves ACL structs into their own file before we add more.
