+++ agent/consul/structs/structs.go
@@ -9,7 +9,6 @@ import (
 	"strings"
 	"time"
 
-	"github.com/hashicorp/consul/acl"
 	"github.com/hashicorp/consul/api"
 	"github.com/hashicorp/consul/types"
 	"github.com/hashicorp/go-msgpack/codec"
@@ -32,6 +31,8 @@ type RaftIndex struct {
 	ModifyIndex uint64
 }
 
+// These are serialized between Consul servers and stored in Consul snapshots,
+// so entries must only ever be added to the end.
 const (
 	RegisterRequestType MessageType = iota
 	DeregisterRequestType
@@ -71,13 +72,6 @@ const (
 
 	// metaValueMaxLength is the maximum allowed length of a metadata value
 	metaValueMaxLength = 512
-
-	// Client tokens have rules applied
-	ACLTypeClient = "client"
-
-	// Management tokens have an always allow policy.
-	// They are used for token management.
-	ACLTypeManagement = "management"
 
 	// MaxLockDelay provides a maximum LockDelay value for
 	// a session. Any value above this will not be respected.
@@ -748,103 +742,6 @@ func (r *SessionSpecificRequest) RequestDatacenter() string {
 type IndexedSessions struct {
 	Sessions Sessions
 	QueryMeta
-}
-
-// ACL is used to represent a token and its rules
-type ACL struct {
-	ID    string
-	Name  string
-	Type  string
-	Rules string
-
-	RaftIndex
-}
-type ACLs []*ACL
-
-type ACLOp string
-
-const (
-	ACLSet      ACLOp = "set"
-	ACLForceSet       = "force-set" // Deprecated, left to backwards compatibility
-	ACLDelete         = "delete"
-)
-
-// IsSame checks if one ACL is the same as another, without looking
-// at the Raft information (that's why we didn't call it IsEqual). This is
-// useful for seeing if an update would be idempotent for all the functional
-// parts of the structure.
-func (a *ACL) IsSame(other *ACL) bool {
-	if a.ID != other.ID ||
-		a.Name != other.Name ||
-		a.Type != other.Type ||
-		a.Rules != other.Rules {
-		return false
-	}
-
-	return true
-}
-
-// ACLRequest is used to create, update or delete an ACL
-type ACLRequest struct {
-	Datacenter string
-	Op         ACLOp
-	ACL        ACL
-	WriteRequest
-}
-
-func (r *ACLRequest) RequestDatacenter() string {
-	return r.Datacenter
-}
-
-// ACLRequests is a list of ACL change requests.
-type ACLRequests []*ACLRequest
-
-// ACLSpecificRequest is used to request an ACL by ID
-type ACLSpecificRequest struct {
-	Datacenter string
-	ACL        string
-	QueryOptions
-}
-
-func (r *ACLSpecificRequest) RequestDatacenter() string {
-	return r.Datacenter
-}
-
-// ACLPolicyRequest is used to request an ACL by ID, conditionally
-// filtering on an ID
-type ACLPolicyRequest struct {
-	Datacenter string
-	ACL        string
-	ETag       string
-	QueryOptions
-}
-
-func (r *ACLPolicyRequest) RequestDatacenter() string {
-	return r.Datacenter
-}
-
-type IndexedACLs struct {
-	ACLs ACLs
-	QueryMeta
-}
-
-type ACLPolicy struct {
-	ETag   string
-	Parent string
-	Policy *acl.Policy
-	TTL    time.Duration
-	QueryMeta
-}
-
-// ACLReplicationStatus provides information about the health of the ACL
-// replication system.
-type ACLReplicationStatus struct {
-	Enabled          bool
-	Running          bool
-	SourceDatacenter string
-	ReplicatedIndex  uint64
-	LastSuccess      time.Time
-	LastError        time.Time
 }
 
 // Coordinate stores a node name with its associated network coordinate.
